+++ packages/eslint-config-sheriff/src/handpickedRules/vitestHandPickedRules.ts
@@ -1,3 +1,5 @@
+import type { TSESLint } from '@typescript-eslint/utils';
+
 export const vitestHandPickedRules = {
   'vitest/consistent-test-it': [2, { fn: 'test', withinDescribe: 'test' }],
   'vitest/expect-expect': 2,
@@ -34,4 +36,4 @@ export const vitestHandPickedRules = {
   'vitest/require-top-level-describe': 2,
   'vitest/valid-describe-callback': 2,
   'vitest/valid-expect': 2,
-};
+} as const satisfies TSESLint.FlatConfig.Rules;