+++ packages/eslint-config-sheriff/src/utils/getIndexedBaseNoRestrictedSyntaxRules.ts
@@ -1,10 +1,12 @@
 import { baseNoRestrictedSyntaxRules } from '@sherifforg/constants';
+import type { NoRestrictedSyntaxSlice } from '@sherifforg/types';
 
-export const getIndexedBaseNoRestrictedSyntaxRules = () => {
-  return baseNoRestrictedSyntaxRules.map((element, index) => {
-    return {
-      ...element,
-      message: `${element.message}\n\nIndex: ${index.toString()}.`,
-    };
-  });
-};
+export const getIndexedBaseNoRestrictedSyntaxRules =
+  (): NoRestrictedSyntaxSlice[] => {
+    return baseNoRestrictedSyntaxRules.map((element, index) => {
+      return {
+        ...element,
+        message: `${element.message}\n\nIndex: ${index.toString()}.`,
+      };
+    });
+  };