+++ packages/sheriff-cli/eslint.config.js
@@ -1,5 +1,4 @@
-import sheriff from 'eslint-config-sheriff';
-import { defineFlatConfig } from 'eslint-define-config';
+import { sheriff, tseslint } from 'eslint-config-sheriff';
 
 const sheriffOptions = {
   react: false,
@@ -12,11 +11,14 @@ const sheriffOptions = {
   vitest: false,
 };
 
-export default defineFlatConfig([
-  ...sheriff(sheriffOptions),
+export default tseslint.config(
+  sheriff(sheriffOptions),
   {
     rules: {
       'fsecond/prefer-destructured-optionals': 0,
     },
   },
-]);
+  {
+    ignores: ['**/tsup.config.js'],
+  },
+);