+++ packages/sheriff-cli/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sherifforg/cli
 
+## 8.1.0
+
+### Minor Changes
+
+- 823696e: feat(config): replace eslint-define-config with tseslint.config().
+  Fixes #343
+
+## 8.0.0
+
+### Major Changes
+
+- 3e6fc65: feat(cli): implemented @sherifforg/cli
+  Closes [#332](https://github.com/AndreaPontrandolfo/sheriff/issues/332)
+
 ## 7.2.0
 
 ### Minor Changes